[Refactor] Header 프로젝트 구조에 맞게 변경
