+++ src/layout/Header/components/index.js
@@ -1,0 +1,2 @@
+export { default as MainHeader } from './MainHeader.jsx';
+export { default as SubHeader } from './SubHeader.jsx';
