+++ src/layout/Header/ui/style.js
@@ -1,0 +1,8 @@
+import {styled} from "styled-components";
+
+export const Container = styled.header`
+  position: sticky;
+  width: 100%;
+  background-color: ${(props) => props.theme.colors.white};
+  box-shadow: 0 0 29px rgba(0,0,0,0.13);
+`
