+++ src/layout/RootLayout.jsx
@@ -1,5 +1,5 @@
 import {Outlet} from 'react-router-dom';
-import Header from '@layout/Header/Header.jsx';
+import Header from '@layout/Header/ui/Header.jsx';
 import Footer from '@layout/Footer/Footer.jsx';
 
 const RootLayout = () => {
